READ ME update
This is my first website designed using the Bootstrap 3 framework.  It is completely responsive for all pages and features a carousel on the home page and a responsive toggle navbar throughout the site.  Please note that the comics page does take some time to load because of the amount of pictures (will dd more on other pages) and content for the Video Game and Collectable pages will be coming soon.